shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-shared"
version = "0.1.0"
description = "Shared utilities for Strands agents"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "colorlog>=6.8.0",
    "pydantic>=2.0.0",
    "colorama>=0.4.6",
    "strands-agents>=1.16.0",
    "boto3>=1.40.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
]